requirements.txt
django==4.2.3
djangorestframework==3.14.0
markdown
django-filter
psycopg2-binary==2.9.6
celery==5.3.1
redis==4.6.0
python-dotenv==1.0.0
drf-spectacular==0.26.3
whitenoise==5.2.0
django-celery-beat
django-celery-results
requests==2.31.0
celery==5.3.1
